% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_dodd.R
\name{adherence_dodd}
\alias{adherence_dodd}
\title{Estimate a Participant's Probability of Adhering to "N Days Per Week" Type of 
Physical Activity Guideline (Dodd's Method)}
\usage{
adherence_dodd(n, x, n.rec = 7, x.rec = 5, posterior = NULL)
}
\arguments{
\item{n}{Number of monitoring days.}

\item{x}{Number of active days.}

\item{n.rec}{Denominator for recommendation.}

\item{x.rec}{Numerator for recommendation.}

\item{posterior}{Can be \code{NULL} for original Dodd method or 
\code{"mean"} or \code{"median"} for modified version described above.}
}
\description{
Implements the Bayesian approach developed by Dodd and used in the landmark 
Troiano et al. paper (\emph{MSSE} 2008).
}
\details{
The approach aims to estimate a participant's probability of meeting 
guidelines of the form "at least x minutes per day for at least y days per 
week" based on observing X active days out of n monitoring days. We 
illustrate here with the "5+ active days per week" guideline that motivated 
the approach.

The prior assumption for the participant's \emph{daily} adherence probability 
is:

p_d ~ Uni(0, 1)

Given p_d, the number of active days out of n monitoring days is distributed:

X|p_d ~ Bin(n, p_d)

It can be shown that the posterior for p_d is:

p_d|X ~ Beta(X + 1, n - X + 1)

Under a somewhat questionable independence assumption, the weekly adherence 
probability is p_w = P(Y >= 5) with Y ~ Bin(7, p_d). Dodd estimates p_w as:

p_w.hat = P(p_d >= 5/7 | X)

which can be calculated using \code{\link[stats:Beta]{pbeta}}.

In my view, the quantity P(p_d >= 5/7 | X) is not a good estimator for p_w. 
Consider what would happen in a really long protocol. The Beta posterior for 
p_d would be very tightly centered around the true p_d, and 
p_w.hat = P(p_d >= 5/7 | X) would be very close to either 0 or 1 -- not very 
close to what we're trying to estimate, p_w.

A solution is to define p_d.hat as the posterior mean, median, or mode, and
map that estimate to p_w, i.e. p_w.hat = P(Y >= 5) with Y ~ Bin(7, p_d.hat). 
So there is an option for that.
}
\examples{
# Generate data from hypothetical study with 1000 subjects, valid days 
# randomly sampled from 1-7, and p_d's drawn from Beta(0.5, 3).
set.seed(1)
n <- sample(1: 7, size = 1000, replace = TRUE)
p_d <- rbeta(n = 1000, shape1 = 0.5, shape2 = 3)
x <- rbinom(n = 1000, size = n, prob = p_d)

# Estimate p_w's using Dodd's method
p_w.hat <- adherence_dodd(n = n, x = x)

# Note that the mean p_w.hat differs considerably from the true mean p_w, 
# reflecting bias in the estimator.
mean(p_w.hat)
mean(pbinom(q = 4, size = 7, prob = p_d, lower.tail = FALSE))


}
\references{
Dodd, K. (2008). Estimation of the population prevalence of adherence to 
physical activity recommendations based on NHANES accelerometry measurements. 
Technical Report. Available at: 
\url{https://epi.grants.cancer.gov/nhanes_pam/bayesian_adherence_estimation.pdf}. 
Accessed Nov. 13, 2018.

Troiano, R.P., Berrigan, D., Dodd, K.W., Masse, L.C. and McDowell, M. (2008). 
Physical activity in the United States measured by accelerometer. Medicine \& 
Science in Sports \& Exercise 40(1): 181--188.
}
