% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_garriguet.R
\name{adherence_garriguet}
\alias{adherence_garriguet}
\title{Estimate a Participant's Probability of Adhering to "N Days Per Week" Type of 
Physical Activity Guideline (Garriguet's Method)}
\usage{
adherence_garriguet(n = 7, x, goal = 5, alpha, beta)
}
\arguments{
\item{n}{Number of monitoring days.}

\item{x}{Number of exercise days.}

\item{goal}{Recommended number of active days out of \code{n}.}

\item{alpha}{Parameter in p_d ~ Beta(alpha, beta). Corresponds to 
\code{shape1} in \code{\link[stats]{Beta}} functions.}

\item{beta}{Parameter in p_d ~ Beta(alpha, beta). Corresponds to 
\code{shape2} in \code{\link[stats]{Beta}} functions.}
}
\description{
Implements the Bayesian approach described by Garriguet 
(\emph{Statistics Canada} 2016).
}
\details{
The approach aims to estimate a participant's probability of meeting 
guidelines of the form "at least x minutes per day for at least y days per 
week" based on observing X active days out of n monitoring days.

The prior assumption for the participant's \emph{daily} adherence probability 
is:

p_d ~ Beta(alpha, beta)

where alpha and beta are estimated via maximum likelihood using the observed 
sample proportions if active days for all study participants. This can be 
done separately via \code{\link{mles_beta}}.

Given p_d, the number of active days out of n monitoring days is distributed:

X|p_d ~ Bin(n, p_d)

It can be shown that the posterior for p_d is:

p_d|X ~ Beta(alpha2 = alpha + X, beta2 = beta + n - X)

Garriguet then uses the Beta-binomial distribution, which describes binomial 
data with success probability randomly drawn from Beta(alpha, beta). The 
weekly adherence estimator is defined as:

p_w.hat <- P(Y >= 5) with Y ~ Betabin(7, alpha2, beta2)

which can be calculated using \code{\link{mles_beta}}.
}
\examples{
# Generate data from hypothetical study with 100 subjects, valid days 
# randomly sampled from 1-7, and p_d's drawn from Beta(1, 2).
set.seed(1)
n <- sample(1: 7, size = 100, replace = TRUE)
p_d <- rbeta(n = 100, shape1 = 1, shape2 = 2)
x <- rbinom(n = 100, size = n, prob = p_d)

# First step: Estimate (alpha, beta) via maximum likelihood
mles <- fitdistr(x = x / n, densfun = "beta", start = list(shape1 = 1, shape2 = 1))



}
\references{
Garriguet, D. (2016). Using a betabinomial distribution to estimate the 
prevalence of adherence to physical activity guidelines among children and 
youth. Statistics Canada, Catalogue no. 82-003-X. Health Reports 27(4): 3-9. 
Available at: \url{https://www150.statcan.gc.ca/n1/pub/82-003-x/2016004/article/14489-eng.pdf}.
}
