% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mles_beta.R
\name{mles_beta}
\alias{mles_beta}
\title{Get Maximum Likelihood Estimates for Beta Distribution}
\usage{
mles_beta(x, start = c(0.5, 0.5))
}
\arguments{
\item{x}{Observations assumed to be iid Beta(alpha, beta).}

\item{start}{Starting values for alpha and beta.}
}
\description{
Same idea as \code{fitdistr} function in \strong{MASS}, but has default 
starting values and uses \code{\link[stats]{nlminb}} rather than 
\code{\link[stats]{optim}}.
}
\examples{
# Generate data from Beta(1, 2) and get MLE's
set.seed(1)
x <- rbeta(n = 1000, shape1 = 1, shape2 = 2)
mles <- mles_beta(x)
mles$par

}
