% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerometry-data.R
\docType{data}
\name{tridata}
\alias{tridata}
\title{Triaxial Sample Data}
\description{
Toy dataset with triaxial minute-to-minute counts generated from a trivariate 
normal distribution. Does not closely resemble real accelerometer data.
}
