% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accelerometry-data.R
\docType{data}
\name{unidata}
\alias{unidata}
\title{Uniaxial Sample Data}
\source{
\url{https://wwwn.cdc.gov/nchs/nhanes/search/datapage.aspx?Component=Examination&CycleBeginYear=2003}
}
\description{
Accelerometer data for the first 5 participants in the National Health and 
Nutrition Examination Survey (NHANES) 2003-2004 dataset.
}
